+++ src/api/tutors/router.py
@@ -3,6 +3,8 @@ from typing_extensions import Annotated
 from fastapi import APIRouter, UploadFile, Depends, status, Query
 from sqlalchemy.orm import Session
 
+from src.api.auth.jwt import InvalidJwt, JwtResolver, get_jwt_resolver
+from src.api.auth.service import AuthenticationService
 from src.api.exceptions import (
     Duplicated,
     EntityNotFound,
@@ -10,6 +12,7 @@ from src.api.exceptions import (
     InvalidFileType,
     ServerError,
 )
+from src.api.users.exceptions import InvalidCredentials
 from src.api.users.repository import UserRepository
 from src.api.tutors.service import TutorService
 from src.api.tutors.schemas import (
@@ -18,13 +21,12 @@ from src.api.tutors.schemas import (
     TutorResponse,
     TutorList,
     PeriodList,
-    TutorResponseWithTopics,
     TutorWithTopicsList,
 )
 from src.api.tutors.repository import TutorRepository
 from src.api.tutors.exceptions import InvalidPeriod
-
 from src.api.auth.hasher import get_hasher, ShaHasher
+from src.api.auth.schemas import oauth2_scheme
 from src.config.database.database import get_db
 
 router = APIRouter(prefix="/tutors")
@@ -39,6 +41,7 @@ router = APIRouter(prefix="/tutors")
     responses={
         status.HTTP_400_BAD_REQUEST: {"description": "The columns are not correct"},
         status.HTTP_409_CONFLICT: {"description": "Duplicated tutor"},
+        status.HTTP_401_UNAUTHORIZED: {"description": "Invalid token"},
         status.HTTP_415_UNSUPPORTED_MEDIA_TYPE: {
             "description": "Content-Type is not correct."
         },
@@ -50,24 +53,30 @@ async def upload_csv_file(
     file: UploadFile,
     hasher: Annotated[ShaHasher, Depends(get_hasher)],
     session: Annotated[Session, Depends(get_db)],
+    token: Annotated[str, Depends(oauth2_scheme)],
+    jwt_resolver: Annotated[JwtResolver, Depends(get_jwt_resolver)],
     period: str = Query(...),
 ):
     try:
+        auth_service = AuthenticationService(jwt_resolver)
+        auth_service.assert_only_admin(token)
         # Check if content-type is a text/csv
         if file.content_type != "text/csv":
             raise InvalidFileType("CSV file must be provided")
         content = (await file.read()).decode("utf-8")
         service = TutorService(TutorRepository(session))
         res = service.create_tutors_from_csv(
-            content, period, hasher, UserRepository(session))
+            content, period, hasher, UserRepository(session)
+        )
 
         return res
     except (InvalidCsv, EntityNotFound, Duplicated, InvalidFileType) as e:
         raise e
+    except InvalidJwt as e:
+        raise InvalidCredentials("Invalid Authorization")
     except Exception as e:
         raise ServerError(str(e))
-
-
+    
 @router.delete(
     "/{tutor_id}",
     description="Deletes a tutor",
@@ -75,19 +84,29 @@ async def upload_csv_file(
     tags=["Tutors"],
     responses={
         status.HTTP_404_NOT_FOUND: {"description": "Tutor id not found"},
+        status.HTTP_401_UNAUTHORIZED: {"description": "Invalid token"},
         status.HTTP_500_INTERNAL_SERVER_ERROR: {"description": "Internal server error"},
     },
     status_code=status.HTTP_202_ACCEPTED,
 )
-async def delete_tutor(tutor_id: int, session: Annotated[Session, Depends(get_db)]):
+async def delete_tutor(
+    tutor_id: int,
+    session: Annotated[Session, Depends(get_db)],
+    token: Annotated[str, Depends(oauth2_scheme)],
+    jwt_resolver: Annotated[JwtResolver, Depends(get_jwt_resolver)],
+):
     try:
+        auth_service = AuthenticationService(jwt_resolver)
+        auth_service.assert_only_admin(token)
+
         service = TutorService(TutorRepository(session))
         return service.delete_tutor(tutor_id)
     except EntityNotFound as e:
         raise e
+    except InvalidJwt as e:
+        raise InvalidCredentials("Invalid Authorization")
     except Exception as e:
         raise ServerError(str(e))
-
 
 @router.post(
     "/periods",
@@ -97,22 +116,29 @@ async def delete_tutor(tutor_id: int, session: Annotated[Session, Depends(get_db
     tags=["Periods"],
     responses={
         status.HTTP_400_BAD_REQUEST: {"description": "Period schema is not correct"},
+        status.HTTP_401_UNAUTHORIZED: {"description": "Invalid token"},
         status.HTTP_409_CONFLICT: {"description": "Duplicated period"},
         status.HTTP_500_INTERNAL_SERVER_ERROR: {"description": "Internal server error"},
     },
     status_code=status.HTTP_201_CREATED,
 )
 async def add_period(
-    session: Annotated[Session, Depends(get_db)], period: PeriodRequest
+    session: Annotated[Session, Depends(get_db)],
+    period: PeriodRequest,
+    token: Annotated[str, Depends(oauth2_scheme)],
+    jwt_resolver: Annotated[JwtResolver, Depends(get_jwt_resolver)],
 ):
     try:
+        auth_service = AuthenticationService(jwt_resolver)
+        auth_service.assert_only_admin(token)
         service = TutorService(TutorRepository(session))
         return service.add_period(period)
     except (InvalidPeriod, Duplicated) as e:
         raise e
+    except InvalidJwt as e:
+        raise InvalidCredentials("Invalid Authorization")
     except Exception as e:
         raise ServerError(str(e))
-
 
 @router.get(
     "/periods",
@@ -122,20 +148,26 @@ async def add_period(
     tags=["Periods"],
     status_code=status.HTTP_200_OK,
     responses={
+        status.HTTP_401_UNAUTHORIZED: {"description": "Invalid token"},
         status.HTTP_500_INTERNAL_SERVER_ERROR: {"description": "Internal server error"},
     },
 )
 async def get_periods(
     session: Annotated[Session, Depends(get_db)],
+    token: Annotated[str, Depends(oauth2_scheme)],
+    jwt_resolver: Annotated[JwtResolver, Depends(get_jwt_resolver)],
     order: str = Query(pattern="^(ASC|DESC)$", default="DESC"),
 ):
     try:
+        auth_service = AuthenticationService(jwt_resolver)
+        auth_service.assert_only_admin(token)
         service = TutorService(TutorRepository(session))
         periods = service.get_all_periods(order)
         return periods
+    except InvalidJwt as e:
+        raise InvalidCredentials("Invalid Authorization")
     except Exception as e:
         raise ServerError(str(e))
-
 
 @router.post(
     "/{tutor_id}/periods",
@@ -145,24 +177,30 @@ async def get_periods(
     tags=["Periods"],
     status_code=status.HTTP_201_CREATED,
     responses={
-        status.HTTP_409_CONFLICT: {"description": "Duplicated period"},
+        status.HTTP_401_UNAUTHORIZED: {"description": "Invalid token"},
         status.HTTP_404_NOT_FOUND: {"description": "Tutor not found"},
+        status.HTTP_409_CONFLICT: {"description": "Duplicated period"},
         status.HTTP_500_INTERNAL_SERVER_ERROR: {"description": "Internal server error"},
     },
 )
 async def add_period_to_tutor(
     session: Annotated[Session, Depends(get_db)],
     tutor_id: int,
+    token: Annotated[str, Depends(oauth2_scheme)],
+    jwt_resolver: Annotated[JwtResolver, Depends(get_jwt_resolver)],
     period_id: str = Query(...),
 ):
     try:
+        auth_service = AuthenticationService(jwt_resolver)
+        auth_service.assert_only_admin(token)
         service = TutorService(TutorRepository(session))
         return service.add_period_to_tutor(tutor_id, period_id)
     except (Duplicated, EntityNotFound) as e:
         raise e
+    except InvalidJwt as e:
+        raise InvalidCredentials("Invalid Authorization")
     except Exception as e:
         raise ServerError(str(e))
-
 
 @router.get(
     "/{tutor_id}/periods",
@@ -172,6 +210,7 @@ async def add_period_to_tutor(
     tags=["Periods"],
     status_code=status.HTTP_200_OK,
     responses={
+        status.HTTP_401_UNAUTHORIZED: {"description": "Invalid token"},
         status.HTTP_404_NOT_FOUND: {"description": "Tutor not found"},
         status.HTTP_500_INTERNAL_SERVER_ERROR: {"description": "Internal server error"},
     },
@@ -179,15 +218,20 @@ async def add_period_to_tutor(
 async def get_tutor_periods(
     session: Annotated[Session, Depends(get_db)],
     tutor_id: int,
+    token: Annotated[str, Depends(oauth2_scheme)],
+    jwt_resolver: Annotated[JwtResolver, Depends(get_jwt_resolver)],
 ):
     try:
+        auth_service = AuthenticationService(jwt_resolver)
+        auth_service.assert_only_admin(token)
         service = TutorService(TutorRepository(session))
         return service.get_periods_by_tutor_id(tutor_id)
     except EntityNotFound as e:
         raise e
+    except InvalidJwt as e:
+        raise InvalidCredentials("Invalid Authorization")
     except Exception as e:
         raise ServerError(str(e))
-
 
 @router.get(
     "/periods/{period_id}",
@@ -197,6 +241,7 @@ async def get_tutor_periods(
     tags=["Periods"],
     status_code=status.HTTP_200_OK,
     responses={
+        status.HTTP_401_UNAUTHORIZED: {"description": "Invalid token"},
         status.HTTP_404_NOT_FOUND: {"description": "Period not found"},
         status.HTTP_500_INTERNAL_SERVER_ERROR: {"description": "Internal server error"},
     },
@@ -204,11 +249,17 @@ async def get_tutor_periods(
 async def get_tutor_periods(
     session: Annotated[Session, Depends(get_db)],
     period_id: str,
+    token: Annotated[str, Depends(oauth2_scheme)],
+    jwt_resolver: Annotated[JwtResolver, Depends(get_jwt_resolver)],
 ):
     try:
+        auth_service = AuthenticationService(jwt_resolver)
+        auth_service.assert_only_admin(token)
         service = TutorService(TutorRepository(session))
         return service.get_tutors_by_period_id(period_id)
     except EntityNotFound as e:
         raise e
+    except InvalidJwt as e:
+        raise InvalidCredentials("Invalid Authorization")
     except Exception as e:
         raise ServerError(str(e))
